cleanup auto lod tool
