+++ ydr/properties.py
@@ -26,161 +26,6 @@ from bpy.app.handlers import persistent
 from bpy.path import basename
 
 
-class AutoLODLevelSettings(bpy.types.PropertyGroup):
-    """Per-LOD-level settings for auto LOD generation."""
-
-    ratio: FloatProperty(
-        name="Ratio",
-        description="Decimation ratio for this LOD level (1.0 = no reduction, 0.0 = maximum reduction)",
-        min=0.01, max=1.0, default=0.5,
-        subtype="FACTOR"
-    )
-
-    use_target_tri_count: BoolProperty(
-        name="Use Target Tri Count",
-        description="Use a target triangle count instead of a ratio",
-        default=False
-    )
-
-    target_tri_count: IntProperty(
-        name="Target Tris",
-        description="Target number of triangles for this LOD level",
-        min=4, default=1000
-    )
-
-
-class AutoLODSettings(bpy.types.PropertyGroup):
-    """Settings for the auto LOD generation tool."""
-
-    ref_mesh: PointerProperty(
-        type=bpy.types.Mesh,
-        name="Reference Mesh",
-        description="The mesh to copy and decimate for each LOD level. Usually the highest quality LOD"
-    )
-
-    levels: lod_level_enum_flag_prop_factory()
-
-    decimate_method: EnumProperty(
-        name="Method",
-        description="Decimation method to use",
-        items=[
-            ("COLLAPSE", "Collapse", "Edge collapse decimation - best general-purpose method"),
-            ("UNSUBDIV", "Un-Subdivide", "Reverse subdivision - best for meshes created via subdivision"),
-            ("DISSOLVE", "Planar", "Planar decimation - dissolves geometry on flat surfaces"),
-        ],
-        default="COLLAPSE"
-    )
-
-    use_per_lod_ratios: BoolProperty(
-        name="Per-LOD Ratios",
-        description="Use individual settings for each LOD level instead of a uniform step",
-        default=False,
-        update=lambda self, ctx: self._on_per_lod_ratios_changed()
-    )
-
-    decimate_step: FloatProperty(
-        name="Decimate Step",
-        description="Uniform decimation step applied for each LOD level",
-        min=0.0, max=0.99, default=0.6
-    )
-
-    decimate_from_original: BoolProperty(
-        name="Decimate from Original",
-        description="Always decimate from the reference mesh instead of cascading from the previous LOD",
-        default=False
-    )
-
-    lod_very_high: PointerProperty(type=AutoLODLevelSettings)
-    lod_high: PointerProperty(type=AutoLODLevelSettings)
-    lod_medium: PointerProperty(type=AutoLODLevelSettings)
-    lod_low: PointerProperty(type=AutoLODLevelSettings)
-    lod_very_low: PointerProperty(type=AutoLODLevelSettings)
-
-    preserve_uvs: BoolProperty(
-        name="Preserve UVs",
-        description="Preserve UV seam boundaries during decimation",
-        default=True
-    )
-
-    preserve_sharp: BoolProperty(
-        name="Preserve Sharp",
-        description="Preserve edges marked as sharp",
-        default=False
-    )
-
-    preserve_vertex_groups: BoolProperty(
-        name="Preserve Vertex Groups",
-        description="Preserve vertex group boundaries",
-        default=False
-    )
-
-    preserve_materials: BoolProperty(
-        name="Preserve Materials",
-        description="Preserve material slot boundaries",
-        default=False
-    )
-
-    planar_angle_limit: FloatProperty(
-        name="Angle Limit",
-        description="Maximum angle between face normals for planar decimation",
-        min=0.0, max=3.14159,
-        default=0.087266,
-        subtype="ANGLE"
-    )
-
-    unsubdiv_iterations: IntProperty(
-        name="Iterations",
-        description="Number of un-subdivide iterations",
-        min=1, max=10, default=2
-    )
-
-    auto_set_distances: BoolProperty(
-        name="Auto-Set LOD Distances",
-        description="Automatically calculate and set LOD distances on the parent Drawable based on mesh size",
-        default=False
-    )
-
-    auto_merge_materials: BoolProperty(
-        name="Merge Materials",
-        description="Automatically bake and merge all materials into a single material for each generated LOD, using the Material Merge settings",
-        default=False
-    )
-
-    _RECOMMENDED_RATIOS = {
-        LODLevel.VERYHIGH: 0.90,
-        LODLevel.HIGH: 0.70,
-        LODLevel.MEDIUM: 0.50,
-        LODLevel.LOW: 0.30,
-        LODLevel.VERYLOW: 0.15,
-    }
-
-    def _on_per_lod_ratios_changed(self):
-        """Set recommended per-LOD ratios when per-LOD mode is enabled."""
-        if self.use_per_lod_ratios:
-            self.reset_ratios()
-
-    def reset_ratios(self):
-        """Reset per-LOD ratios to recommended defaults."""
-        for lod_level, ratio in self._RECOMMENDED_RATIOS.items():
-            lod_settings = self.get_lod_settings(lod_level)
-            lod_settings.ratio = ratio
-            lod_settings.use_target_tri_count = False
-
-    def get_lod_settings(self, lod_level) -> "AutoLODLevelSettings":
-        """Get the per-LOD settings for a given LODLevel."""
-        match lod_level:
-            case LODLevel.VERYHIGH:
-                return self.lod_very_high
-            case LODLevel.HIGH:
-                return self.lod_high
-            case LODLevel.MEDIUM:
-                return self.lod_medium
-            case LODLevel.LOW:
-                return self.lod_low
-            case LODLevel.VERYLOW:
-                return self.lod_very_low
-
-
 class ShaderOrderItem(bpy.types.PropertyGroup):
     # For drawable shader order list
     index: bpy.props.IntProperty(min=0)
@@ -870,9 +715,6 @@ def register():
     bpy.types.Light.time_flags = bpy.props.PointerProperty(type=LightTimeFlags)
     bpy.types.Light.light_flags = bpy.props.PointerProperty(type=LightFlags)
 
-    bpy.types.Scene.sollumz_auto_lod_settings = bpy.props.PointerProperty(type=AutoLODSettings)
-
-
     bpy.types.Scene.sollumz_extract_lods_levels = lod_level_enum_flag_prop_factory()
     bpy.types.Scene.sollumz_extract_lods_parent_type = bpy.props.EnumProperty(name="Parent Type", items=(
         ("sollumz_extract_lods_parent_type_object", "Object", "Parent to an Object"),
@@ -1013,7 +855,6 @@ def unregister():
     del bpy.types.Scene.create_seperate_drawables
     del bpy.types.Scene.auto_create_embedded_col
     del bpy.types.Scene.center_drawable_to_selection
-    del bpy.types.Scene.sollumz_auto_lod_settings
     del bpy.types.Scene.sollumz_extract_lods_levels
     del bpy.types.Scene.sollumz_extract_lods_parent_type
